added downsampler.ts and downsampler_worker_factory.ts
