+++ src/wasi_snapshot.ts
@@ -1,137 +1,137 @@
 export const wasiSnapshotPreview1Emulator = {
-    args_get: (input: any): any => {
-        return 0;
-    },
-    args_sizes_get: (input: any): any => {
-        return 0;
-    },
-    environ_get: (input: any): any => {
-        return 0;
-    },
-    environ_sizes_get: (input: any): any => {
-        return 0;
-    },
-    clock_res_get: (input: any): any => {
-        return 0;
-    },
-    clock_time_get: (input: any): any => {
-        return 0;
-    },
-    fd_advise: (input: any): any => {
-        return 0;
-    },
-    fd_allocate: (input: any): any => {
-        return 0;
-    },
-    fd_close: (input: any): any => {
-        return 0;
-    },
-    fd_datasync: (input: any): any => {
-        return 0;
-    },
-    fd_fdstat_get: (input: any): any => {
-        return 0;
-    },
-    fd_fdstat_set_flags: (input: any): any => {
-        return 0;
-    },
-    fd_fdstat_set_rights: (input: any): any => {
-        return 0;
-    },
-    fd_filestat_get: (input: any): any => {
-        return 0;
-    },
-    fd_filestat_set_size: (input: any): any => {
-        return 0;
-    },
-    fd_filestat_set_times: (input: any): any => {
-        return 0;
-    },
-    fd_pread: (input: any): any => {
-        return 0;
-    },
-    fd_prestat_get: (input: any): any => {
-        return 0;
-    },
-    fd_prestat_dir_name: (input: any): any => {
-        return 0;
-    },
-    fd_pwrite: (input: any): any => {
-        return 0;
-    },
-    fd_read: (input: any): any => {
-        return 0;
-    },
-    fd_readdir: (input: any): any => {
-        return 0;
-    },
-    fd_renumber: (input: any): any => {
-        return 0;
-    },
-    fd_seek: (input: any): any => {
-        return 0;
-    },
-    fd_sync: (input: any): any => {
-        return 0;
-    },
-    fd_tell: (input: any): any => {
-        return 0;
-    },
-    fd_write: (input: any): any => {
-        return 0;
-    },
-    path_create_directory: (input: any): any => {
-        return 0;
-    },
-    path_filestat_get: (input: any): any => {
-        return 0;
-    },
-    path_filestat_set_times: (input: any): any => {
-        return 0;
-    },
-    path_link: (input: any): any => {
-        return 0;
-    },
-    path_open: (input: any): any => {
-        return 0;
-    },
-    path_readlink: (input: any): any => {
-        return 0;
-    },
-    path_remove_directory: (input: any): any => {
-        return 0;
-    },
-    path_rename: (input: any): any => {
-        return 0;
-    },
-    path_symlink: (input: any): any => {
-        return 0;
-    },
-    path_unlink_file: (input: any): any => {
-        return 0;
-    },
-    poll_oneoff: (input: any): any => {
-        return 0;
-    },
-    proc_exit: (input: any): any => {
-        return 0;
-    },
-    proc_raise: (input: any): any => {
-        return 0;
-    },
-    sched_yield: (input: any): any => {
-        return 0;
-    },
-    random_get: (input: any): any => {
-        return 0;
-    },
-    sock_recv: (input: any): any => {
-        return 0;
-    },
-    sock_send: (input: any): any => {
-        return 0;
-    },
-    sock_shutdown: (input: any): any => {
-        return 0;
-    },
+  args_get: (input: any): any => {
+    return 0;
+  },
+  args_sizes_get: (input: any): any => {
+    return 0;
+  },
+  environ_get: (input: any): any => {
+    return 0;
+  },
+  environ_sizes_get: (input: any): any => {
+    return 0;
+  },
+  clock_res_get: (input: any): any => {
+    return 0;
+  },
+  clock_time_get: (input: any): any => {
+    return 0;
+  },
+  fd_advise: (input: any): any => {
+    return 0;
+  },
+  fd_allocate: (input: any): any => {
+    return 0;
+  },
+  fd_close: (input: any): any => {
+    return 0;
+  },
+  fd_datasync: (input: any): any => {
+    return 0;
+  },
+  fd_fdstat_get: (input: any): any => {
+    return 0;
+  },
+  fd_fdstat_set_flags: (input: any): any => {
+    return 0;
+  },
+  fd_fdstat_set_rights: (input: any): any => {
+    return 0;
+  },
+  fd_filestat_get: (input: any): any => {
+    return 0;
+  },
+  fd_filestat_set_size: (input: any): any => {
+    return 0;
+  },
+  fd_filestat_set_times: (input: any): any => {
+    return 0;
+  },
+  fd_pread: (input: any): any => {
+    return 0;
+  },
+  fd_prestat_get: (input: any): any => {
+    return 0;
+  },
+  fd_prestat_dir_name: (input: any): any => {
+    return 0;
+  },
+  fd_pwrite: (input: any): any => {
+    return 0;
+  },
+  fd_read: (input: any): any => {
+    return 0;
+  },
+  fd_readdir: (input: any): any => {
+    return 0;
+  },
+  fd_renumber: (input: any): any => {
+    return 0;
+  },
+  fd_seek: (input: any): any => {
+    return 0;
+  },
+  fd_sync: (input: any): any => {
+    return 0;
+  },
+  fd_tell: (input: any): any => {
+    return 0;
+  },
+  fd_write: (input: any): any => {
+    return 0;
+  },
+  path_create_directory: (input: any): any => {
+    return 0;
+  },
+  path_filestat_get: (input: any): any => {
+    return 0;
+  },
+  path_filestat_set_times: (input: any): any => {
+    return 0;
+  },
+  path_link: (input: any): any => {
+    return 0;
+  },
+  path_open: (input: any): any => {
+    return 0;
+  },
+  path_readlink: (input: any): any => {
+    return 0;
+  },
+  path_remove_directory: (input: any): any => {
+    return 0;
+  },
+  path_rename: (input: any): any => {
+    return 0;
+  },
+  path_symlink: (input: any): any => {
+    return 0;
+  },
+  path_unlink_file: (input: any): any => {
+    return 0;
+  },
+  poll_oneoff: (input: any): any => {
+    return 0;
+  },
+  proc_exit: (input: any): any => {
+    return 0;
+  },
+  proc_raise: (input: any): any => {
+    return 0;
+  },
+  sched_yield: (input: any): any => {
+    return 0;
+  },
+  random_get: (input: any): any => {
+    return 0;
+  },
+  sock_recv: (input: any): any => {
+    return 0;
+  },
+  sock_send: (input: any): any => {
+    return 0;
+  },
+  sock_shutdown: (input: any): any => {
+    return 0;
+  },
 };
