+++ src/worker_types.ts
@@ -59,22 +59,13 @@ export interface DownsamplingWorker extends Omit<Worker, 'postMessage'> {
   postMessage(command: DownsamplingWorkerRequest): void;
 }
 
-export type DownsamplerWasmOutput = {
-  inputBufferAddress: number,
-  inputframeLength: number,
-  memory: WebAssembly.Memory,
-  objectAddress: number,
-  outputBufferAddress: number,
-  pvDownsamplerConvertNumSamplesToInputSampleRate: CallableFunction,
-  pvDownsamplerInit: CallableFunction,
-  pvDownsamplerProcess: CallableFunction,
-  pvDownsamplerReset: CallableFunction,
-  pvDownsamplerDelete: CallableFunction,
-}
-
 export interface DownsamplerInterface {
   delete(): void;
-  getrequiredNumInputSamples(numSample: number): number;
-  process(inputBuffer: Int16Array, inputBufferSize: number, outputBuffer: Int16Array): number;
+  getNumRequiredInputSamples(numSample: number): number;
+  process(
+    inputBuffer: Int16Array,
+    inputBufferSize: number,
+    outputBuffer: Int16Array,
+  ): number;
   reset(): void;
 }
